requirements.txt
pandas
requests
Pillow
torch
torchvision
